Move all routes to route config
