+++ app/javascript/app/pages/root/root.js
@@ -2,13 +2,13 @@ import { PureComponent, createElement } from 'react';
 import Proptypes from 'prop-types';
 import { connect } from 'react-redux';
 import { actions } from 'providers/countries-provider';
-import Component from './container-component';
+import Component from './root-component';
 
 const mapStateToProps = state => ({
   countriesLoaded: state.countries.loaded
 });
 
-class Container extends PureComponent {
+class Root extends PureComponent {
   constructor(props) {
     super(props);
     props.getCountries();
@@ -21,9 +21,9 @@ class Container extends PureComponent {
   }
 }
 
-Container.propTypes = {
+Root.propTypes = {
   getCountries: Proptypes.func,
   countriesLoaded: Proptypes.bool
 };
 
-export default connect(mapStateToProps, actions)(Container);
+export default connect(mapStateToProps, actions)(Root);
