+++ package.json
@@ -26,8 +26,9 @@
     "react-motion": "^0.5.0",
     "react-player": "0.20.0",
     "react-redux": "^5.0.5",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
+    "react-router": "4.2.0",
+    "react-router-config": "1.0.0-beta.4",
+    "react-router-dom": "4.2.0",
     "react-select": "1.0.0-rc.5",
     "react-simple-maps": "0.7.2",
     "recharts": "^1.0.0-alpha.2",
